Percentage display fix

+++ src/Model.h
@@ -49,7 +49,7 @@ private:
     float windDirection;
 
 	int burnedAreaCounter;
-	int completeArea;
+	float completeArea;
 
     bool setRandomVehicleLocation(ofRectangle area, bool liveInWater, ofVec2f & location);
